mon: Monitor: Add initial support for synchronization (WARNING: WIP)
This is very on-going work, and I've actually stopped working on it for the
time being while addressing other pieces of this puzzle. And I feel I can't
emphasize this enough: this is on-going work.

There is a lot of stuff missing, some other stuff was completely left in
a limbo of half-done/half-undone, and it will certainly NOT WORK.

Also, there are a couple of things that are to be replaced, some other
stuff will be stripped away.

(update as of 08/10/12: the Monitor class compiles; will not work as the
'handle_sync()' function is commented out)

mon: Monitor: messaging for synchronizing two monitor stores
mon: Monitor: ignore heartbeat replies if we are no longer synchronizing
mon: Monitor: backoff trim enable once the last sync requester finishes
mon: Monitor: wait for the leader to reply to 'finish' before bootstrapping
mon: Monitor: clear store before starting fresh sync
paxos: change paxos_max_join_drift comment to sync instead of slurp

Signed-off-by: Joao Eduardo Luis <joao.luis@inktank.com>